+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [v0.0.13]
+
+### Added
+
+- Guard Data Context - use data within the data sources to validate guards.
+
+### Fixed
+
+- Nullable Entities Patched
+
 ## [v0.0.12]
 
 ### Added